book & user struct data type
I have added a struct data type which is used to record details about the book found in library( or when the library add new), I have defined it at the time of declaration because we use this information every where so it should be global!, the array size can be changed. the second struct is to input user (reader) information.